Use the Symfony Routing on Nova's Routing Service
