+++ composer.json
@@ -18,6 +18,7 @@
         "symfony/finder": "^3.1",
         "symfony/http-foundation": "^3.1",
         "symfony/http-kernel": "^3.1",
+        "symfony/routing": "^3.1",
         "stack/builder": "^1.0",
         "doctrine/inflector": "1.0.*",
         "phpmailer/phpmailer": "5.2.*",
